moved the LLVM globals code to a new file LLVMGlobalConst.hs
